pyproject.toml
[project]
name = "obs-twitch-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openai>=1.76.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]
